requirements.txt
APScheduler==3.10.1
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
PyPDF2==3.0.1
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.30.0
six==1.16.0
soupsieve==2.4.1
tzdata==2023.3
tzlocal==4.3
urllib3==2.0.2
